refactor(ucmc-web): consolidate role row sheets into single tabbed editor
Three icon buttons per row crowded the mobile layout. Replaces them with
a single Pencil button that opens one RoleEditorSheet with tabs for
Members / Permissions / Edit. Shared dirty-state guard covers both
write tabs at once. Drops the three single-purpose dialog files now
that they're inlined as tab panels.
